Extract Creator Kit routes out of agent-channel.ts (Wave D batch 5)
agent-channel.ts 2771 -> 2643 lines. Moves KIT/KIT_API into a new
agent-channel-kit.ts as registerAgentChannelKitRoutes(app, deps),
same pattern as batches 1-4. Completes the kit trio started by batch 1's
agent-channel-kit-files.ts (KIT -> KIT_API -> KIT_FILES).

store and objectStore are passed by reference (| undefined, matching
the established pattern) since both are shared with other routes still
in the parent file. gateVerdict -- a thin closure over
options.gamesStore, called from channelState/KIT/GATE -- is passed by
reference the same way resolveBuild already is. KIT_ENTRY/
KitRegistryError/kitUnpackCommand/parseKitRegistry/parseKitSidecar/
compactKitDigestForApi/DEFAULT_MCP_DIGEST_MAX_BYTES move with the
cluster (their only callers).
