+++ eslint-rules/module-boundary-map.mjs
@@ -178,6 +178,7 @@ const FILE_BUCKET = {
   'agent-channel-examples': 'agent-surface',
   'agent-channel-brief': 'agent-surface',
   'agent-channel-seed': 'agent-surface',
+  'agent-channel-kit': 'agent-surface',
   'mcp-server': 'agent-surface',
   'mcp-tool-support': 'agent-surface',
   'mcp-example-tools': 'agent-surface',
